Start work splitting out residential search controller for AP01 and CH01
