+++ src/controllers/update/update.director.residential.address.search.controller.ts
@@ -1,0 +1,12 @@
+import { getDirectorResidentialAddressSearch, postDirectorResidentialAddressSearch } from "../shared.controllers/director.residential.address.search.controller"
+import { NextFunction, Request, Response } from "express"
+import { UPDATE_DIRECTOR_RESIDENTIAL_ADDRESS_PATH } from "../../types/page.urls"
+import { Templates } from "../../types/template.paths"
+
+export const get = (req: Request, resp: Response, next: NextFunction) => {
+  getDirectorResidentialAddressSearch(req, resp, next, Templates.UPDATE_DIRECTOR_RESIDENTIAL_ADDRESS_SEARCH, UPDATE_DIRECTOR_RESIDENTIAL_ADDRESS_PATH);
+}
+
+export const post = async (req: Request, resp: Response, next: NextFunction) => {
+  postDirectorResidentialAddressSearch(req, resp, next, Templates.UPDATE_DIRECTOR_RESIDENTIAL_ADDRESS_SEARCH, UPDATE_DIRECTOR_RESIDENTIAL_ADDRESS_PATH, true);
+};
